refactor handlers to directory structure
